Endponits - Recipe
Implementación de los endponits
buscar por nombre
buscar entre fechas
buscar por nutricionist_id

+++ src/test/java/com/appnutricare/service/RecipeServiceImplTest.java
@@ -97,7 +97,7 @@ public class RecipeServiceImplTest {
 
     @Test
     void findAllByNutritionist() throws Exception {
-        Long nutritionistId = 1L;
+        Integer nutritionistId = 1;
         List<Recipe> recipeList;
 
         ProfessionalProfile professionalProfile = new ProfessionalProfile(1, "description1");
